@@ -11,7 +11,6 @@
         "@mui/icons-material": "^5.11.16",
         "@mui/material": "^5.12.3",
         "@mui/styled-engine-sc": "^5.12.0",
-        "@mui/styles": "^5.13.2",
         "@mui/system": "^5.13.2",
         "@mui/x-data-grid": "^5.17.26",
         "@mui/x-date-pickers": "^5.0.20",
@@ -36,19 +35,19 @@
         "jest-junit": "^16.0.0",
         "module-federation-import-remote": "^1.0.18",
         "npm": "^9.6.7",
-        "react": "17.0.2",
+        "react": "18.2.0",
         "react-cookie": "^6.1.1",
         "react-dev-utils": "^12.0.0",
-        "react-dom": "17.0.2",
+        "react-dom": "18.2.0",
         "react-final-form": "6.5.9",
-        "react-final-form-listeners": "^1.0.3",
+        "react-final-form-arrays": "^3.1.4",
         "react-imask": "^6.4.2",
         "react-number-format": "^4.9.3",
         "react-redux": "7.2.9",
         "react-refresh": "^0.11.0",
         "react-router-dom": "5.3.4",
         "react-text-mask": "^5.4.3",
-        "react-xml-viewer": "^1.3.0",
+        "react-xml-viewer": "^2.0.1",
         "redux": "4.2.1",
         "redux-thunk": "2.4.2",
         "semantic-release": "^21.1.2",
@@ -63,28 +62,25 @@
 
     },
     "devDependencies": {
-        "@babel/core": "7.17.5",
-        "@babel/plugin-proposal-class-properties": "^7.12.1",
-        "@babel/plugin-proposal-object-rest-spread": "^7.17.3",
+        "@babel/core": "^7.17.5",
+        "@babel/plugin-transform-class-properties": "^7.23.3",
+        "@babel/plugin-transform-object-rest-spread": "^7.23.4",
         "@babel/plugin-syntax-jsx": "^7.16.7",
         "@babel/plugin-transform-async-to-generator": "^7.16.8",
         "@babel/plugin-transform-runtime": "^7.17.0",
         "@babel/preset-env": "^7.12.1",
         "@babel/preset-react": "7.16.7",
         "@emotion/jest": "^11.11.0",
-        "@testing-library/jest-dom": "^6.1.4",
-        "@testing-library/react": "^12.1.5",
-        "@testing-library/react-hooks": "^8.0.1",
-        "@wojtekmaj/enzyme-adapter-react-17": "^0.8.0",
+        "@testing-library/jest-dom": "^6.1.5",
+        "@testing-library/react": "^13.0.0",
         "autosuggest-highlight": "^3.3.4",
-        "babel-eslint": "^10.1.0",
+        "@babel/eslint-parser": "^7.23.3",
         "babel-loader": "8.2.3",
         "clean-webpack-plugin": "^3.0.0",
         "copy-webpack-plugin": "^6.2.1",
         "cross-env": "^7.0.3",
         "css-loader": "^6.7.1",
-        "enzyme": "^3.11.0",
-        "enzyme-adapter-react-17-updated": "^1.0.2",
+        "external-remotes-plugin": "1.0.0",
         "eslint": "^8.55.0",
         "eslint-config-prettier": "^9.1.0",
         "eslint-plugin-import": "^2.29.0",
@@ -92,20 +88,18 @@
         "eslint-plugin-prettier": "^5.0.1",
         "eslint-plugin-react": "^7.33.2",
         "eslint-plugin-sonarjs": "^0.23.0",
-        "external-remotes-plugin": "1.0.0",
-        "html-webpack-plugin": "5.5.0",
         "husky": "^8.0.3",
+        "html-webpack-plugin": "5.5.0",
+        "node-sass": "^9.0.0",
         "jest": "^29.7.0",
         "jest-environment-jsdom": "^29.7.0",
-        "jest-enzyme": "^7.1.2",
         "jest-sonar-reporter": "^2.0.0",
         "jest-transform-stub": "^2.0.0",
-        "node-sass": "^6.0.1",
         "prettier": "^3.1.1",
         "prop-types": "^15.8.1",
-        "sass-loader": "^10.0.3",
+        "sass-loader": "^13.3.2",
         "style-loader": "^2.0.0",
-        "webpack": "5.70.0",
+        "webpack": "^5.76.0",
         "webpack-cli": "4.10.0",
         "webpack-dev-server": "v4.8.0",
         "webpack-livereload-plugin": "3.0.2",
